upload

+++ src/main/java/com/example/microfinancepi/services/OfferLoanServiceImpl.java
@@ -1,13 +1,18 @@
 package com.example.microfinancepi.services;
 
-import com.example.microfinancepi.entities.OfferLoan;
-import com.example.microfinancepi.entities.RequestLoan;
+import com.example.microfinancepi.entities.*;
 import com.example.microfinancepi.repositories.OfferLoanRepository;
 import com.example.microfinancepi.repositories.RequestLoanRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
 
 @Service
@@ -17,7 +22,43 @@ public class OfferLoanServiceImpl implements  IOfferLoanService {
 
     OfferLoanRepository offerLoanRepo;
     RequestLoanRepository requestLoanRepo;
+    UserRepository userRepository;
+    private User_role getCurrentUserRole(User user) {
+        return user.getRole();
+    }
+    private final String TMM_URL = "https://www.stb.com.tn/fr/site/bourse-change/historique-des-tmm/";
 
+    @Scheduled(cron = "0 0 9 1 * ?") // Run at 9 AM on the first day of every month
+    public void updateTmmValues() {
+        Double tmmValue = fetchLastTmmValue(TMM_URL);
+        if (tmmValue != null) {
+            List<OfferLoan> offerLoans = offerLoanRepo.findAll();
+            for (OfferLoan offerLoan : offerLoans) {
+                offerLoan.setTmm(tmmValue);
+                offerLoanRepo.save(offerLoan);
+            }
+            log.info("TMM values updated successfully for {} offer loans.", offerLoans.size());
+        } else {
+            log.error("Failed to fetch TMM value. Updating TMM values aborted.");
+        }
+    }
+    public Double fetchLastTmmValue(String url) {
+        Double lastTmmValue = null;
+        try {
+            final Document doc = Jsoup.connect(url).get();
+            for (Element row : doc.select("table.cours-de-change tr")) {
+                if (row.select("td:nth-of-type(1)").text().equals("")) {
+                    continue;
+                } else {
+                    final String rate = row.select("td:nth-of-type(3)").text();
+                    lastTmmValue = Double.parseDouble(rate);
+                }
+            }
+        } catch (Exception e) {
+            log.error("Error fetching TMM values: {}", e.getMessage());
+        }
+        return lastTmmValue;
+    }
     public List<OfferLoan> retrieveAllOfferLoans() {
         return offerLoanRepo.findAll();
     }
@@ -26,38 +67,74 @@ public class OfferLoanServiceImpl implements  IOfferLoanService {
         return offerLoanRepo.findById(idOffer).get();
     }
 
-    public OfferLoan addOfferLoan(OfferLoan o) {
-        return offerLoanRepo.save(o);
-        //if (!user.getRoles().contains(RoleName.ADMIN)) {
-        //      throw new AccessDeniedException("you're not authorized");
-        //    }
-    }
+    public OfferLoan addOfferLoan(OfferLoan offer, User user) {
+        User_role userRole = getCurrentUserRole(user);
 
-    public OfferLoan modifyOfferLoan(OfferLoan offer) {
+        if (!Arrays.asList(User_role.ADMIN, User_role.AGENT).contains(userRole)) {
+            throw new AccessDeniedException("You're not authorized to add an offer");
+        }
+        Double lastTmmValue = fetchLastTmmValue("https://www.stb.com.tn/fr/site/bourse-change/historique-des-tmm/");
+        offer.setTmm(lastTmmValue);
+
+        offer.setUser(user);
         return offerLoanRepo.save(offer);
     }
 
-    public void removeOfferLoan(Long idOffer) {
+    public OfferLoan modifyOfferLoan(OfferLoan offer, User user) {
+        User_role userRole = getCurrentUserRole(user);
+        if (!Arrays.asList(User_role.ADMIN, User_role.AGENT).contains(userRole)) {
+            throw new AccessDeniedException("You're not authorized to add an offer");
+        }
+        offer.setUser(user);
+        return offerLoanRepo.save(offer);
+    }
+
+    public void removeOfferLoan(Long idOffer, User user) {
+        if (!user.getRole().equals(User_role.ADMIN)) {
+            throw new AccessDeniedException("Only admins are authorized to delete offers.");
+        }
         offerLoanRepo.deleteById(idOffer);
     }
 
-    // Assign + unassign child to/from parent
-    public void assignRequestToOffer(Long idOffer, Long idReqt) {
-        OfferLoan offerLoan = offerLoanRepo.findById(idOffer).get();//parent
-        RequestLoan requestLoan = requestLoanRepo.findById(idReqt).get();//child
-
-        offerLoan.getRequestloans().add(requestLoan);
-        offerLoanRepo.save(offerLoan);
+    public float simulationCalMonthlyRepaymentAmount(Long offerId, Long loanAmnt, Long nbrMonth) {
+        OfferLoan offer = offerLoanRepo.findById(offerId).get();
+        LoanType loanType = offer.getTypeLoan();
+        if (loanType == LoanType.STUDENT || loanType == LoanType.FARMER || loanType == LoanType.HOMEMAKER) {
+            float intRate = offer.getIntRate();
+            float annualIntRate = intRate / 100;
+            float monthlyIntRate= annualIntRate/12;
+            return (float) ((loanAmnt * monthlyIntRate) / (1 - Math.pow(1 + monthlyIntRate, -nbrMonth*12)));
+        } else {
+            throw new IllegalArgumentException("Constant Monthly repayment calculation is only applicable for LoanType STUDENT, FARMER, or HOMEMAKER.");
+        }
     }
 
-    public void unassignRequestFromOffer(Long idOffr, Long idReqt) {
-        OfferLoan offerLoan = offerLoanRepo.findById(idOffr).get();//parent
-        RequestLoan requestLoan = requestLoanRepo.findById(idReqt).get();//child
+    public float simulationCalYearlyRepaymentAmount(Long offerId, Long loanAmnt, Long nbrYears, typeAmort type) {
+        OfferLoan offer = offerLoanRepo.findById(offerId).get();
 
-        offerLoan.getRequestloans().remove(requestLoan);
-        offerLoanRepo.save(offerLoan);
+        if (offer.getTypeLoan() != LoanType.PROJECT) {
+            throw new IllegalArgumentException("Yearly repayment calculation is only applicable for LoanType PROJECT.");
+        }
+        float intRate = offer.getIntRate();
+        float annualRate = intRate / 100;
+        switch (type) {
+            case CONST_ANNUITY:
+                return (float) ((loanAmnt * annualRate) / (1 - Math.pow(1 + annualRate, -nbrYears)));
+            case CONST_AMORTIZATION:
+                float amort = loanAmnt / nbrYears;
+                float interest = loanAmnt * annualRate;
+                return amort + interest;
+            case LOAN_PER_BLOC:
+                float sumInterest = 0;
+                for (int i = 0; i < nbrYears; i++) {
+                    sumInterest += loanAmnt * annualRate;
+                }
+                float totalAmount = loanAmnt + sumInterest;
+                return totalAmount;
+            default:
+                throw new IllegalArgumentException("Unsupported amortization type: " + type);
+        }
     }
-
     //search and display offers that the minAmount is greater than the searchedAmount
     public List<OfferLoan> findOffersByMinAmntLessThanEqual(Long searchedAmount) {
         return offerLoanRepo.findOffersByMinAmntLessThanEqual(searchedAmount);
@@ -67,127 +144,29 @@ public class OfferLoanServiceImpl implements  IOfferLoanService {
     public int countAvailableOffers() {
         return offerLoanRepo.countByStatus("AVAILABLE");
     }
+
     public int countRequestLoansByOfferId(Long offerId) {
         return offerLoanRepo.countRequestLoansByOfferId(offerId);
     }
 
+
 }
+
+
     /*
-    //scheduler only in service to display on console
-    @Scheduled(fixedDelay = 15000) // en milliseconde
-    public List<OfferLoan> retrieveAllOffers(){
-
-        List<OfferLoan> listOffr = offerLoanRepo.findAll();
-        log.info("how many offers found:" + listOffr.size());
-        for (OfferLoan offer: listOffr){
-            log.info("list of offers:" + offer);
-        }
-        return  listOffr;
+    // Assign + unassign child to/from parent:
+    public void unassignRequestFromOffer(Long idOffr, Long idReqt) {
+        OfferLoan offerLoan = offerLoanRepo.findById(idOffr).get();//parent
+        RequestLoan requestLoan = requestLoanRepo.findById(idReqt).get();//child
+        offerLoan.getRequestloans().remove(requestLoan);
+        offerLoanRepo.save(offerLoan);
     }
-
-}
-
-methodes wassim:
-@Override
-    public Map<Integer, Map<String, Double>> calculateCapitalAmortiAndRestant(Integer id_offer, Integer id_request) {
-        Offers_Credit offreCredit = Ioffers_creditrepository.findById(id_offer).orElse(null);
-        Request request = IRequestRepository.findById(id_request).orElse(null);
-        if (offreCredit == null || request == null) {
-            throw new IllegalArgumentException("Invalid offre-credit or request id");
-        }
-
-        Map<Integer, Map<String, Double>> result = new LinkedHashMap<>();
-        double capitalRestant = request.getAmount_req();
-
-        double annuity = calculateAnnuity(request.getAmount_req(), offreCredit.getInterest_rate(), Integer.parseInt(request.getTerm_loan()));
-        double term_loan = Integer.parseInt(request.getTerm_loan());
-        double interest_rate = Double.parseDouble(offreCredit.getInterest_rate());
-        double amount_req = request.getAmount_req();
+*/
+//  @Scheduled(cron = "0 */5 * * * *")
+    /*public void scheduledFetchAndUpdateTmmValue() {
+        fetchAndUpdateTmmValue();
 
 
-        for (int year = 1; year <= Integer.parseInt(request.getTerm_loan()); year++) {
-            Map<String, Double> yearData = new LinkedHashMap<>();
-            double capitalAmorti = calculateCapitalAmorti(capitalRestant, annuity, offreCredit.getInterest_rate());
-            capitalRestant -= capitalAmorti;
-            yearData.put("capitalAmorti", capitalAmorti);
-            yearData.put("capitalRestant", capitalRestant);
-            yearData.put("term_loan", term_loan);
-            yearData.put("interest_rate", interest_rate);
-            yearData.put("amount_req", amount_req);
-
-            result.put(year, yearData);
-        }
+*/
 
 
-        return result;
-    }
-
-
-    private double calculateAnnuity(double amount, String interestRate, int term) {
-        double rate = Double.parseDouble(interestRate.replace("%", "")) / 100.0;
-        double r = rate / 12.0;
-        double n = term * 12.0;
-        return (amount * r) / (1 - Math.pow(1 + r, -n));
-    }
-
-    private double calculateCapitalAmorti(double capitalRestant, double annuity, String interestRate) {
-        double rate = Double.parseDouble(interestRate.replace("%", "")) / 100.0;
-        double r = rate / 12.0;
-        return annuity - (capitalRestant * r);
-
-    }
-
-    @Override
-    public List<Offers_Credit> findMatchingOffer(Integer amount, String repaymentPeriode) {
-        // Récupérer toutes les offres de crédit
-        List<Offers_Credit> all = Ioffers_creditrepository.findAll();
-
-        List<Offers_Credit> matchingOffers = new ArrayList<>();
-
-        // Parcourir chaque offre de crédit
-        for (Offers_Credit offer : all) {
-            if (amount >= offer.getMin_amount() && amount <= offer.getMax_amount()) {
-                if (repaymentPeriode.equals(offer.getRepayment_period())) {
-
-                    matchingOffers.add(offer);
-
-
-                }
-            }
-        }
-
-        return matchingOffers;
-    }
-   @Override
-    public OfferStatistics getOfferStatistics(Integer id_offer) {
-
-
-      List<Request> requests=IRequestRepository.findRequestsByOfferId(id_offer);
-        Integer acceptedRequests = 0;
-        Integer rejectedRequests = 0;
-        Integer pendingRequests = 0;
-        for (Request request : requests) {
-
-            if (request.getAccOrRef() != null) {
-                if ("Request accepté".equalsIgnoreCase(request.getAccOrRef().getCheck_loan())) {
-                    acceptedRequests++;
-                } else if ("Request réfusé".equalsIgnoreCase(request.getAccOrRef().getCheck_loan())) {
-                    rejectedRequests++;
-                }
-            } else {
-                pendingRequests++;
-            }
-        }
-        OfferStatistics statistics = new OfferStatistics();
-        statistics.setOfferId(id_offer);
-        statistics.setAcceptedRequests(acceptedRequests);
-        statistics.setRejectedRequests(rejectedRequests);
-        statistics.setPendingRequests(pendingRequests);
-        return statistics;
-
-
-
-   }
-
-
-     */
